netfilter: flowtable: use the device forward path to get the offload device
When offloading a flow that runs through a bridge and/or vlans, the flow
created for software offloading can often stop at a virtual interface (usually
the bridge). When attempting to offload the flow to hardware, traverse
the forward path again to fill in the information for the real underlying
ethernet device.

Signed-off-by: Felix Fietkau <nbd@nbd.name>